Update Vector.h
Changed it so that size() will return type size_t instead of int

+++ Unit04-Mechanics-of-lists/homework4c-vector/Vector.h
@@ -2,6 +2,7 @@
 
 #include <iostream>
 #include <string>
+#include <cstddef>
 
 template<class T>
 class Vector {
@@ -26,7 +27,7 @@ public:
         // implement operator[] here
     }
 
-    int size() const {
+    size_t size() const {
         // implement size here
     }
